pyproject.toml
[project]
name = "tigr-contract-workspace"
version = "0.1.0"
description = "Workspace tooling for the tigr contract repository"
requires-python = ">=3.11"
dependencies = [
  "pydantic>=2.8",
  "PyYAML>=6.0",
  "jsonschema>=4.23",
  "pytest>=8.3",
]

[tool.pytest.ini_options]
pythonpath = [
  ".",
  "packages/contract-py/src",
  "packages/artifacts-py/src",
]
testpaths = ["tests"]
